% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_print.R
\name{bmatrix}
\alias{bmatrix}
\title{Inline latex method for matrix}
\usage{
bmatrix(x)
}
\arguments{
\item{x}{R matrix object}
}
\value{
This function is designed to use in the latex block. Output in the chunk might useless. For chunk usage, `paste()` of last line should be replaced by `writeLines()`.
}
\description{
Write R matrix object in Latex block.
This function might be useful as an alternative to implementing the whole package.
By this one function, same things can be done as \code{\link{knit_print.matrix}}, althogh you should specify the function.
}
\examples{
A <- matrix(1:9, nrow = 3)
bmatrix(A)
## You can write the matrix A in the inline $`r bmatrix(A)`$ or
## $$`r bmatrix(A)`$$
}
\seealso{
\code{\link{knit_print.matrix}}
}
