% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqn_ref.R
\name{eqn_numbering}
\alias{eqn_numbering}
\title{Equation reference in html documents}
\usage{
eqn_numbering()
}
\value{
Copy and paste to the document
}
\description{
By writing this script at the beginning of the rmd file, equation referencing becomes more flexible in html documents.
}
\examples{
eqn_numbering()
}
