% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_figure.R
\docType{data}
\name{fig}
\alias{fig}
\title{Cross reference a figure caption}
\format{An object of class \code{list} of length 2.}
\usage{
fig
}
\arguments{
\item{refer}{of `$cap` refer the figure caption, same as the chunk name}

\item{title}{of `$cap` caption of the plot}

\item{refer}{of `$ref` cross refer the corresponding figure by number}
}
\value{
The number of the figure increases in order. fig$ref("chunk") returns the figure number followed by the string Figure. You don't have to worry changed order of plots while documentation.
Also, they provide hyperlink in pdf document so that you can easily approach the plots.
}
\description{
list, cross reference a `fig.cap` in r markdown with label and link of pdf
}
\examples{
```{r example-plot, fig.cap = fig$cap("example-plot", "Example plot")}
set.seed(1)
mydf <- data_frame(x = rnorm(100))
mydf <- mydf \%>\% mutate(y = x * 2)
mydf \%>\%
  ggplot(aes(x, y)) +
  geom_point()
```
You might cross reference with latex block $`r fig$ref("example-plot")`$
}
\keyword{datasets}
