% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_print.R
\name{knit_print.tex}
\alias{knit_print.tex}
\alias{knit_print.matrix}
\title{Print tex or matrix class as latex}
\usage{
\method{knit_print}{tex}(x, inline = FALSE, ...)

\method{knit_print}{matrix}(x, inline = FALSE, ...)
}
\arguments{
\item{x}{`tex` or `matrix` to be printed}

\item{inline}{Choose between inline versus in-chunk. The default is set to be `inline = FALSE`, in-chunk.}

\item{...}{Additional arguments passed to the S3 method.}
}
\description{
`knitr_print()` will print an object with math form in the r markdown chunk or inline. It can be an tex or a matrix. In the chunk, `results = 'asis'` option should be assigned to that chunk.
Printing matrix becomes more compact than before. It will be printed math form in the inline, i.e. `r matrix`. Here, the function does not have to be written.
If in-chunk option is choosed, normal matrix form in R console will be printed.
}
\examples{
A <- matrix(1:9, nrow = 3)
library(knitr) # to use knit_print() function directly, knitr package needed
knit_print(as_tex(A))
knit_print(as_tex(A), inline = TRUE)
knit_print(A)
knit_print(A, inline = TRUE)
## In case of matrix, $`r A`$ or $$`r A`$$ works.
## vignette will provide the details.
}
\seealso{
\code{\link{bmatrix}} Printing matrix latex form by using the function
}
