% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_print.R
\name{knitr_print.matrix}
\alias{knitr_print.matrix}
\title{Print matrix as latex}
\usage{
knitr_print.matrix(x, inline = FALSE, ...)
}
\arguments{
\item{x}{`matrix` to be printed}

\item{inline}{Choose between inline versus in-chunk. The default is set to be `inline = FALSE`, in-chunk.}
}
\description{
`knitr_print()` will now print math form in the inline, i.e. `r matrix`. Here, the function does not have to be specified. If in-chunk option is choosed, normal matrix form in R console will be printed.
}
