% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{rmdtool-package}
\alias{rmdtool-package}
\alias{rmdtool}
\title{Private tools to document in R-markdown}
\description{
This is an assistant tool for R Markdown, mainly pdf
}
\author{
Young-geun Kim <\url{https://github.com/ygeunkim}>
}
