% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_table.R
\docType{data}
\name{tab}
\alias{tab}
\title{Cross reference a table caption}
\format{An object of class \code{list} of length 2.}
\usage{
tab
}
\value{
Everything is same but Figure versus Table.
\item{refer}{of `$cap` refer the table caption, same as the chunk name}
\item{title}{of `$cap` caption of the table}
\item{refer}{of `$ref` cross refer the corresponding table by number}
}
\description{
list, cross reference a table in r markdown with label and link of pdf
}
\examples{
knitr::kable(ggplot2::mpg, caption = tab$cap("tab-chunk-name", "Caption name"))
tab$ref("tab-chunk-name")
## For detailed usage, see the vignette
}
\seealso{
\code{\link{fig}} Figure numbering
}
\keyword{datasets}
