% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_print.R
\name{bmatrix}
\alias{bmatrix}
\title{Inline latex method for matrix}
\usage{
bmatrix(x)
}
\arguments{
\item{x}{R matrix object}
}
\value{
This function is designed to use in the latex block. Output in the chunk might useless. For chunk usage, `paste()` of last line should be replaced by `writeLines()`.
}
\description{
Write R matrix object in Latex block
}
\examples{
```{r}
A <- matrix(1:10, nrow = 3)
```
You can write the matrix A in the inline $`r bmatrix(A)`$ or
$$`r bmatrix(A)`$$
}
