% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_figure.R
\docType{data}
\name{fig}
\alias{fig}
\title{Cross reference a figure caption}
\format{An object of class \code{list} of length 2.}
\usage{
fig
}
\value{
The number of the figure increases in order. fig$ref("chunk") returns the figure number followed by the string Figure. You don't have to worry changed order of plots while documentation.
Also, they provide hyperlink in pdf document so that you can easily approach the plots.
\item{refer}{of `$cap` refer the figure caption, same as the chunk name}
\item{title}{of `$cap` caption of the plot}
\item{refer}{of `$ref` cross refer the corresponding figure by number}
}
\description{
list, cross reference a `fig.cap` in r markdown with label and link of pdf
}
\examples{
library(ggplot2)
ggplot(data = mpg, aes(x = displ, y = hwy)) +
  geom_point() +
  labs(caption = fig$cap("fig-chunk-name", "Caption name"))
fig$ref("fig-chunk-name")
## For detailed usage, see the vignette
}
\seealso{
\code{\link{tab}} Table numbering
}
\keyword{datasets}
