% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_print.R
\name{knit_print.tex}
\alias{knit_print.tex}
\alias{knit_print.matrix}
\title{Print tex or matrix class as latex}
\usage{
\method{knit_print}{tex}(x, inline = FALSE, ...)

\method{knit_print}{matrix}(x, inline = FALSE, ...)
}
\arguments{
\item{x}{`tex` or `matrix` to be printed}

\item{inline}{Choose between inline versus in-chunk. The default is set to be `inline = FALSE`, in-chunk.}
}
\description{
`knitr_print()` will print matrix with math form in the r markdown chunk or inline. In the chunk, `results = 'asis'` option should be assigned to that chunk.
Printing matrix becomes more compact. It will be printed math form in the inline, i.e. `r matrix`. Here, the function does not have to be specified. If in-chunk option is choosed, normal matrix form in R console will be printed.
}
\examples{
```{r, results = 'asis'}
A <- matrix(1:10, nrow = 2)
knitr_print(as_tex(A))
```
}
