% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_table.R
\docType{data}
\name{tab}
\alias{tab}
\title{Cross reference a figure caption}
\format{An object of class \code{list} of length 2.}
\usage{
tab
}
\arguments{
\item{refer}{of `$cap` refer the table caption, same as the table name}

\item{title}{of `$cap` caption of the table}

\item{refer}{of `$ref` cross refer the corresponding table by number}
}
\value{
Everything is same but Figure versus Table.
}
\description{
list, cross reference a table in r markdown with label and link of pdf
}
\seealso{
\code{\link{fig}} Figure numbering
}
\keyword{datasets}
